backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
networkx==3.2.1
aiohttp==3.9.1

# OCR相关依赖
paddleocr>=2.7.0
opencv-python>=4.8.0
pillow>=10.0.0

# 数据库连接驱动
pymysql>=1.1.0
psycopg2-binary>=2.9.9
pymssql>=2.2.10

# 测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
